Don't round Lux to improve night lighting reports (#7)
All of the sensor's integration settings allow for some resolution
beneath 1 Lux. The longest integration time has about 66 levels that it
can report between 0 and 1 Lux. Removing the `round/1` lets this be
returned.

This changes the starting Lux values to `0.0` so that floats are always
returned to users for Lux and the filtered Lux.